refactor(bot): extract order-lifecycle services into OrderMonitorMixin (phase 5)
Fourth cut of the AuramaurBot split. The order monitor (pending-fill
finalization + TTL/expiry), orphaned-pending-trade reconciliation, and the
resolution checker move verbatim into auramaur/bot_order_monitor.py as
OrderMonitorMixin. Methods still operate on the bot's self; run() (in bot.py)
spawns these loops via self, and cross-mixin calls (_clear_exit_suppression,
_check_kill_switch) resolve through the shared self.

AST sweep flagged 5 symbols — all local-variable annotations
(ExchangeClient/MarketDiscovery/PaperTrader/PolymarketClient/ResolutionTracker),
which PEP 526 never evaluates at runtime, so no NameError risk; added them under
TYPE_CHECKING for cleanliness.

bot.py: 2218 -> 1856 lines (3819 -> 1856 across the four phase-5 service cuts,
-51%). The remaining bot.py is the orchestrator core — composition
(_init_components), the scheduler (run), and __init__/shutdown — which
appropriately stay. Full suite: 837 passed.
